PSXPC_N X64 patches

+++ SPEC_PSXPC_N/GPU.C
@@ -31,19 +31,19 @@ unsigned long GadwPolygonBuffers[52260];
 
 void GPU_UseOrderingTables(unsigned long* pBuffers, int nOTSize)//5DF68(<), 5F1C8(<)
 {
-	db.order_table[0] = (unsigned long*)((unsigned long)&pBuffers[0]);
-	db.order_table[1] = (unsigned long*)((unsigned long)&pBuffers[nOTSize]);
+	db.order_table[0] = (uintptr_t*)((uintptr_t)&pBuffers[0]);
+	db.order_table[1] = (uintptr_t*)((uintptr_t)&pBuffers[nOTSize]);
 	db.nOTSize = nOTSize;
-	db.pickup_order_table[0] = (unsigned long*)((unsigned long)&GadwOrderingTables_V2[0]);
-	db.pickup_order_table[1] = (unsigned long*)((unsigned long)&GadwOrderingTables_V2[256]);
+	db.pickup_order_table[0] = (uintptr_t*)((uintptr_t)&GadwOrderingTables_V2[0]);
+	db.pickup_order_table[1] = (uintptr_t*)((uintptr_t)&GadwOrderingTables_V2[256]);
 	return;
 }
 
 void GPU_UsePolygonBuffers(unsigned long* pBuffers, int nPBSize)//5DFB0(<), 
 {
 	db.nPBSize = nPBSize;
-	db.poly_buffer[0] = (unsigned long*)((unsigned long)&pBuffers[0]);
-	db.poly_buffer[1] = (unsigned long*)((unsigned long)&pBuffers[nPBSize]);
+	db.poly_buffer[0] = (unsigned long*)((uintptr_t)&pBuffers[0]);
+	db.poly_buffer[1] = (unsigned long*)((uintptr_t)&pBuffers[nPBSize]);
 	return;
 }
